# api.yaml
swagger: "2.0"
info:
  title: Student CRUD API
  description: This is my Student CRUD API Documentation
  version: 1.0
  contact:
    email: nitugiri02@gmail.com
  liscence:
    name: Apache 2.0
    url: "http://www.apache.org/liscences/LIS"
  servers:
    -url: http://localhost:8080
    description: THis is my local Server
    -url: http://example.com/
    description: This is my original URL
schemes: [http]
...
paths:
...
  /post/{id}:
    put:
      summary: "To Updat Student"
      description: "Update Student by ID"
      operationId: "updateStudent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "update student by ID"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Student object"
        required: true
        schema:
          $ref: "#/definitions/Student"
      responses:
        "500":
          description: "Internal server error"
        "404":
          description: "Student not found"
      x-swagger-router-controller: "Student"
...
definitions:
  Student:
    type: "object"
    properties:
      id:
        type: integer
      name:
        type: "string"
      age:
        type: integer
      roleNo:
          type: integer
      address:
        type: "string"